Move gut of code to addons directory. Support placing api key in ENV
