+++ addon/initializers/aviary-service.js
@@ -1,0 +1,6 @@
+export default {
+  name: 'aviary-service',
+  initialize: function(container, application) {
+    application.inject('component:aviary-editor', 'aviaryApiClient', 'service:aviary-api-client');
+  }
+};
